Better error checking and docs for BW listener tests.
Some of the BW tests will fail if the is only a single STREAM_BW
event available. This happens when the test stream completes in
under 1 second.

This commits adds an exception to graceful skip the part of the
tests which calculates BW_SAMPLES when not enough measurements
are available.