Fix AlreadyCalled exception after a request is canceled
The HTTP request over the circuit can be canceled when it takes
too long by calling the `readBody.cancel()` method.

This will close the client connection immediately. However the
`_readBodyProtcol.connectionLost()` method will also try and call
the `readBody` Deferred errback when it detects the connection has
closed.

As the `readBody` Deferred has already been called this caused
a `AlreadyCalled` exception to be raised. To avoid this
the `cancelableReadBodyProtocol` checks if the Deferred has been
called before calling the `errback` chain.