Use Deferred semaphore to limit simultaneous measurements
The bandwidth scanner was trying to launch all of its Tor measurement
circuits as soon as possible. This will cause issues where the bandwidth
and CPU limits of the measurement machine will distort the measured
bandwidth values for the measured relays.

This commit adds a deferred semaphore and a command line option to
limit the number of simultanious connections. The defaults is 10 but
this should be changed if anyone has a better value.

Resolves #24.