Compare fetched bandwidth files by hash digest
It is not necessary to cache all the bandwidth measurement files for
each fetch in memory. Instead we can just keep a running hash of the
bytes we fetch and use the hash digest to confirm we received the
file successfully.

This commit replaces the `readBodyProtocol` with a
`hashingReadBodyProtocol` which fires a deferred with the SHA1 hex
digest of the response body when the download completes.

Resolves #32.