Prevent authorities from being selected in the exit position
Authorities should never be measured or selected as exits
(#20797). This change fixes a bug where authorities with
the Exit flag but a `Reject *:*` exit policy were being
selected as Exits. This caused random circuit failures
in the tests.

The complete solution involves also reading the Exit policy to
decide if a relay can exit to the bwauth server.